styleai/requirements.txt
flask
opencv-python
pillow
numpy
python-dotenv
groq